refactor: extract the definitions to header file
