+++ src/oruguru.c
@@ -1,14 +1,39 @@
+#include "oruguru.h"
 #include <raylib.h>
 #include <raymath.h>
 #include <stdbool.h>
 
-enum InputType {
-  INPUT_UP = 1 << 0,
-  INPUT_LEFT = 1 << 1,
-  INPUT_DOWN = 1 << 2,
-  INPUT_RIGHT = 1 << 3,
-  INPUT_SELECT = 1 << 4,
-};
+/* PROTOTYPES */
+void draw_inputs(int input_flags);
+void get_inputs(int *input_flags);
+
+/* FUNCTIONS */
+int main(int argc, char *argv[]) {
+  const int screenWidth = 800;
+  const int screenHeight = 450;
+  int input_flags;
+
+  InitWindow(screenWidth, screenHeight, "oruguru");
+  SetTargetFPS(60);
+
+  while (!WindowShouldClose()) {
+    // Update
+    get_inputs(&input_flags);
+
+    // Draw
+    BeginDrawing();
+    {
+      ClearBackground(RAYWHITE);
+      DrawText("oruguru", 190, 200, 20, LIGHTGRAY);
+
+      draw_inputs(input_flags);
+    }
+    EndDrawing();
+  }
+
+  CloseWindow();
+  return 0;
+}
 
 void get_inputs(int *input_flags) {
   *input_flags = 0;
@@ -38,31 +63,4 @@ void draw_inputs(int input_flags) {
     color = input_flags & (1 << i) ? RED : BLUE;
     DrawCircle(x + radius + 2 * radius * i, y + radius, radius, color);
   }
-}
-
-int main(int argc, char *argv[]) {
-  const int screenWidth = 800;
-  const int screenHeight = 450;
-  int input_flags;
-
-  InitWindow(screenWidth, screenHeight, "oruguru");
-  SetTargetFPS(60);
-
-  while (!WindowShouldClose()) {
-    // Update
-    get_inputs(&input_flags);
-
-    // Draw
-    BeginDrawing();
-    {
-      ClearBackground(RAYWHITE);
-      DrawText("oruguru", 190, 200, 20, LIGHTGRAY);
-
-      draw_inputs(input_flags);
-    }
-    EndDrawing();
-  }
-
-  CloseWindow();
-  return 0;
 }
