+++ src/oruguru.h
@@ -1,0 +1,12 @@
+#ifndef ORUGURU_H
+#define ORUGURU_H
+
+enum InputType {
+  INPUT_UP = 1 << 0,
+  INPUT_LEFT = 1 << 1,
+  INPUT_DOWN = 1 << 2,
+  INPUT_RIGHT = 1 << 3,
+  INPUT_SELECT = 1 << 4,
+};
+
+#endif // !ORUGURU_H
